chore: split the deserialization logic into separate functions
