+++ src/deserialization/consts.ts
@@ -1,0 +1,14 @@
+export const commonPushOPMax = 76;
+// Hex
+export const zeroByte = "00";
+export const versionOne = "01000000";
+export const versionTwo = "02000000";
+export const markerFlag = "0001";
+export const coinbaseTXID =
+  "0000000000000000000000000000000000000000000000000000000000000000";
+export const coinbaseVOUT = "ffffffff";
+
+export enum BTC_ENV {
+  MAINNET = "MAINNET",
+  TESTNET = "TESTNET",
+}
